requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
docopt==0.6.2
idna==3.3
pipreqs==0.4.11
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
qrcode==7.3.1
requests==2.26.0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
yarg==0.1.9
